weather-ml/requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary
requests
pandas
scikit-learn
joblib
numpy
